Complete route, add middleware in route in user service

+++ User service/controllers/addressController.js
@@ -5,16 +5,34 @@ const getAllAddress = async (req, res, next) => {
     try {
         let addressList = await Address.find()
         if(!addressList){
-            res.status(400).json({
+            return res.status(400).json({
+                success: false,
                 message: "Cannot find address"
             })
         }else{
-            res.status(500).json({
-                message: `Error: ${error.message}`
+            return res.status(200).json({
+                success: true,
+                message: `Get all address`,
+                addresses: addressList
             })
         }
     } catch (error) {
-        res.status(500).json({
+        return res.status(500).json({
+            message: `Error: ${error.message}`
+        })
+    }
+}
+const getAddressCurrentUser = async (req, res, next) => {
+    try {
+        let user = req.user
+        let addressList = await Address.find({user_id: user.user_id})
+        return res.status(200).json({
+            success: true,
+            message: `Address of current user`,
+            address: addressList
+        })
+    } catch (error) {
+        return res.status(500).json({
             message: `Error: ${error.message}`
         })
     }
@@ -24,25 +42,28 @@ const getAddressByUserId = async (req, res, next) => {
     try {
         let userId = req.params.userId
         if(!userId){
-            res.status(400).json({
+            return res.status(400).json({
+                success: false,
                 message: `user ID is required`
             })
         }else{
             let user = User.findOne({user_id:userId})
             if(!user){
-                res.status(400).json({
+                return res.status(400).json({
+                    success: false,
                     message: `Cannot find user with ID ${userId}`
                 })
             }else{
                 let addressList = await Address.find({user_id: user.user_id})
-                res.status(200).json({
+                return res.status(200).json({
+                    success: true,
                     message: `Address of user ID ${userId}`,
                     address: addressList
                 })
             }
         }
-    } catch (error) {
-        res.status(500).json({
+    }catch(error) {
+        return res.status(500).json({
             message: `Error: ${error.message}`
         })
     }
@@ -51,25 +72,28 @@ const getAddressByUserId = async (req, res, next) => {
 const getAddressById = async (req, res, next) => {
     try {
         let addressId = req.params.addressId
-        if(addressId){
-            res.status(400).json({
+        if(!addressId){
+            return res.status(400).json({
+                success: false,
                 message: `Address ID is required`
             })
         }else{
             let address = Address.findOne({address_id:addressId})
             if(!address){
-                res.status(400).json({
+                return res.status(400).json({
+                    success: false,
                     message: `Cannot find user with ID ${addressId}`
                 })
             }else{
-                res.status(200).json({
+                return res.status(200).json({
+                    success: true,
                     message: `Address ID ${addressId}`,
                     address: address
                 })
             }
         }
     } catch (error) {
-        res.status(500).json({
+        return res.status(500).json({
             message: `Error: ${error.message}`
         })
     }
@@ -77,62 +101,61 @@ const getAddressById = async (req, res, next) => {
 
 const createAddress = async (req,res,next) => {
     try {
-        let userId = req.params.userId
-        let user = User.findOne({user_id:userId})
-        if(!user){
-            res.status(400).json({
-                message: `Cannot find user with ID ${userId}`
+        let user = req.user
+        let city = req.body.city
+        if(!city){
+            return res.status(400).json({
+                success: false,
+                message: "City is required"
+            })
+        }
+        let district = req.body.district
+        if(!district){
+            return res.status(400).json({
+                success: false,
+                message: "District is required"
+            })
+        }
+        let avenue = req.body.avenue
+        if(!avenue){
+            return res.status(400).json({
+                success: false,
+                message: "Avenue is required"
+            })
+        }
+        let specific = req.body.specific
+        if(!specific){
+            return res.status(400).json({
+                success: false,
+                message: "Specific address is required"
+            })
+        }
+        let addressCount = await Address.countDocuments({ user_id: user.userId})
+        let addressId = user.userId + "Address" + (addressCount + 1)
+        let address = new Address({
+            address_id: addressId,
+            user_id: user.userId,
+            city: city,
+            district: district,
+            avenue: avenue,
+            specific: specific
+        })
+
+        let save = await address.save()
+        if(save){
+            return res.status(200).json({
+                success: true,
+                message: "Address saved!",
+                address: address
             })
         }else{
-            let city = req.body.city
-            if(!city){
-                res.status(400).json({
-                    message: "City is required"
-                })
-            }
-            let district = req.body.district
-            if(!district){
-                res.status(400).json({
-                    message: "District is required"
-                })
-            }
-            let avenue = req.body.avenue
-            if(!avenue){
-                res.status(400).json({
-                    message: "Avenue is required"
-                })
-            }
-            let specific = req.body.specific
-            if(!specific){
-                res.status(400).json({
-                    message: "Specific address is required"
-                })
-            }
-            let addressCount = await Address.countDocuments({ user_id: userId})
-            let addressId = userId + "Address" + (addressCount + 1)
-            let address = new Address({
-                address_id: addressId,
-                user_id: userId,
-                city: city,
-                district: district,
-                avenue: avenue,
-                specific: specific
+            return res.status(400).json({
+                success: false,
+                message: "Address save failed!"
             })
-
-            let save = await address.save()
-            if(save){
-                res.status(200).json({
-                    message: "Address saved!",
-                    address: address
-                })
-            }else{
-                res.status(400).json({
-                    message: "Address save failed!"
-                })
-            }
         }
     } catch (error) {
-        res.status(500).json({
+        return res.status(500).json({
             message: `Error: ${error.message}`
         })
     }
@@ -140,19 +163,27 @@ const createAddress = async (req,res,next) => {
 
 const updateAddressById = async (req, res, next) =>{
     try {
+        let user = req.user
         let addressId = req.params.addressId
         if(!addressId){
-            res.status(400).json({
+            return res.status(400).json({
+                success: false,
                 message: "Address Id is required!"
             })
         }else{
-            let address = Address.findOne({address_id: addressId})
+            let address = await Address.findOne({address_id: addressId})
             if(!address){
-                res.status(400).json({
+                return res.status(400).json({
+                    success: false,
                     message: `Cannot find address with ID ${addressId}`
                 })
-            }else{
-                let oldAddress = address
+            }else {
+                if(address.user_id != user.userId){
+                    return res.status(400).json({
+                        success: false,
+                        message: `This address is not yours!`
+                    })
+                }
                 let province = req.body.province
                 if(province){
                     address.province = province
@@ -172,16 +203,16 @@ const updateAddressById = async (req, res, next) =>{
 
                 let save = await address.save()
                 if(save){
-                    res.status(200).json({
+                    return res.status(200).json({
+                        success: true,
                         message: "Address updated!",
-                        before: oldAddress,
-                        after: address
+                        address: address
                     })
                 }
             }
         }
     } catch (error) {
-        res.status(500).json({
+        return res.status(500).json({
             message: `Error: ${error.message}`
         })
     }
@@ -192,30 +223,41 @@ const deleteAddressById = async (req, res, next) => {
     try {
         let addressId = req.params.addressId
     if(!addressId){
-        res.status(400).json({
+        return res.status(400).json({
+            success: false,
             message: "Address ID is required!"
         })
     }else{
+        let user = req.user
         let address = await Address.findOne({address_id: addressId})
         if(!address){
-            res.status(400).json({
+            return res.status(400).json({
+                success: false,
                 message: `Cannot find address with ID ${addressId}`
             })
         }else{
+            if(address.user_id != user.userId){
+                return res.status(400).json({
+                    success: false,
+                    message: `This address is not yours!`
+                })
+            }
             let deleteAddress = await Address.deleteOne({address_id: addressId})
             if(deleteAddress){
-                res.status(200).json({
-                    message: `Deleted address with ID ${addressId}`
+                return res.status(200).json({
+                    success: true,
+                    message: `Deleted address`
                 })
             }else{
-                res.status(400).json({
+                return res.status(400).json({
+                    success: false,
                     message: `Deleted address failed`
                 })
             }
         }
     }
     } catch (error) {
-        res.status(500).json({
+        return res.status(500).json({
             message: `Error: ${error.message}`
         })
     }
@@ -223,6 +265,7 @@ const deleteAddressById = async (req, res, next) => {
 
 module.exports = {
     getAllAddress,
+    getAddressCurrentUser,
     getAddressByUserId,
     getAddressById,
     createAddress,
